CLOSED - # 532: Create utility to import measurements from irouter logs https://verticon-trac.sourcerepo.com/verticon_Tracker/ticket/532

+++ tracker/trunk/com.verticon.tracker/src/com/verticon/tracker/Animal.java
@@ -319,7 +319,7 @@ public interface Animal extends EObject {
 	 * @model transient="true" changeable="false" volatile="true" derived="true"
 	 * @generated
 	 */
-	Integer getWeight();
+	Double getWeight();
 
 	/**
 	 * Returns the value of the '<em><b>Weight Gain Per Day</b></em>' attribute.
